refactor(tui): move pure log renderer into agent_tui_kit
Move render_log_panel_pure (+ its helpers restamp_log_left_border and
render_loading_spinner) into agent_tui_kit::render::log, making the pure
Phase 3 render pass reusable by any ratatui host. The tui crate keeps only
the App-coupled prepare phase (prepare_log_frame) and the two wrappers.

- kit: new render/log.rs (pub render_log_panel_pure; private helpers)
- tui: log.rs drops the moved phase, imports render_log_panel_pure from kit
- tui: cells/mod.rs drops the now-unused code/text/tool re-exports
- tui: code_overlay_tests imports render_code_cards directly from the kit

Pure phase references only kit modules; verify via fmt/test/clippy gate.
